requirements.txt
fastapi
uvicorn[standard]
opencv-python
numpy
ultralytics
paho-mqtt
firebase-admin
google-cloud-firestore
google-cloud-storage
python-dotenv
asyncio
